+++ NEWS.md
@@ -1,8 +1,13 @@
+# iemisc 0.6.2
+
+* Updated these functions: `size`, `righttri`, `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`
+
+
 # iemisc 0.5.1
 
 * Renamed lengths to length_octave
 
-* Updated these functions: `size` `righttri`, `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`
+* Updated these functions: `size`, `righttri`, `Manningtri`, `Manningtrap`, `Manningrect`, `Manningpara`, `Manningcirc`
 
 * Added `iemiscdata` as a suggested R package